backend/requirements.txt
django
djangorestframework
stripe
djangorestframework-simplejwt
django-cors-headers
